+++ rigflow-client/src/ui/mod.rs
@@ -1,7 +1,4 @@
 pub mod app;
-mod app_center;
-mod app_dialogs;
-mod app_panels;
 pub mod bands;
 pub mod frequency_view;
 pub mod layout;
@@ -11,3 +8,8 @@ pub mod state;
 pub mod stats;
 pub mod waterfall;
 pub mod spectrum_view;
+
+mod app_actions;
+mod app_center;
+mod app_dialogs;
+mod app_panels;
